+++ src/Framework/ClassResolver/DocBlockService/UseBlockParser.php
@@ -28,7 +28,7 @@ final class UseBlockParser
 
         $lines = array_filter(
             explode(PHP_EOL, $phpCode),
-            static fn (string $l) => strpos($l, 'use ') === 0 && str_contains($l, $needle)
+            static fn (string $l) => strncmp($l, 'use ', 4) === 0 && str_contains($l, $needle)
         );
 
         /** @psalm-suppress RedundantCast */
@@ -41,7 +41,7 @@ final class UseBlockParser
     {
         $lines = array_filter(
             explode(PHP_EOL, $phpCode),
-            static fn (string $l) => strpos($l, 'namespace ') === 0
+            static fn (string $l) => strncmp($l, 'namespace ', 10) === 0
         );
         /** @psalm-suppress RedundantCast */
         $lineSplit = (array)explode(' ', (string)reset($lines));