+++ src/Framework/FacadeResolverAwareTrait.php
@@ -6,6 +6,10 @@ namespace Gacela\Framework;
 
 use Gacela\Framework\ClassResolver\Facade\FacadeResolver;
 
+/**
+ * @deprecated the new DocBlockResolverAwareTrait is capable of doing the same as FacadeResolverAwareTrait and more
+ * @see DocBlockResolverAwareTrait
+ */
 trait FacadeResolverAwareTrait
 {
     private ?AbstractFacade $facade = null;
@@ -21,6 +25,9 @@ trait FacadeResolverAwareTrait
 
     /**
      * @return class-string
+     *
+     * @deprecated using DocBlockResolverAwareTrait you don't need to define this method anymore
+     * @see DocBlockResolverAwareTrait
      */
     abstract protected function facadeClass(): string;
 }