+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 - Added DocBlockResolverAwareTrait.
+- Deprecated FacadeResolverAwareTrait in favor of DocBlockResolverAwareTrait.
 - Removed deprecated setup as array in Gacela::bootstrap()
 
 ### 0.16.0